novena: Rename sata_boot to sata_root
The sata_boot command is used by extboot, and is always set.  We use
sata_root to determine where the root device is.

This collision causes U-Boot to always attempt to boot from SATA, even
when the root filesystem is set to MMC.

Rename sata_boot to sata_root to work around the collision.

Signed-off-by: Sean Cross <xobs@kosagi.com>